fix lesson_01

+++ lesson_01/area.rb
@@ -2,10 +2,10 @@
 # coding: utf-8
 
 print "Введите основание треугольника: "
-a = gets.chomp.to_i
+a = gets.chomp.to_f
 
 print "Введите высоту треугольника: "
-h = gets.chomp.to_i
+h = gets.chomp.to_f
 
 area = 1 / 2.0 * a * h
 puts "Площадь данного треугольника равна #{area}."
